requirements.txt
pandas>=2.1.0
requests>=2.31.0
pillow>=10.0.0
pydantic>=2.7.0
volcenginesdkarkruntime>=0.0.0